working with path insert
