+++ bundled/scripts/noConfigScripts/debugpy
@@ -1,0 +1,4 @@
+#! /bin/bash
+# Bash script
+python $BUNDLED_DEBUGPY_PATH --listen 0 --wait-for-client $@
+echo "Executed: python $BUNDLED_DEBUGPY_PATH --listen 0 --wait-for-client $@"
