+++ bundled/scripts/noConfigScripts/debugpy.bat
@@ -1,0 +1,3 @@
+@echo off
+:: Bat script
+python %BUNDLED_DEBUGPY_PATH% --listen 0 --wait-for-client %*
