+++ bundled/scripts/noConfigScripts/debugpy.fish
@@ -1,0 +1,2 @@
+# Fish script
+python $BUNDLED_DEBUGPY_PATH --listen 0 --wait-for-client $argv
