+++ bundled/scripts/noConfigScripts/debugpy.ps1
@@ -1,0 +1,2 @@
+# PowerShell script
+python $env:BUNDLED_DEBUGPY_PATH --listen 0 --wait-for-client $args
